backend first draft of authorization

+++ seeders/20200718111517-admin-user.js
@@ -1,4 +1,7 @@
-'use strict';
+const bcrypt = require('bcryptjs');
+const config = require('config');
+
+const SALTROUNDS =  config.get('saltRounds')
 
 module.exports = {
   up: async (queryInterface, Sequelize) => {
@@ -12,11 +15,21 @@ module.exports = {
      * }], {});
     */
 
-   await queryInterface.bulkInsert('Users', [{
-    email: 'siafin2010@gmail.com',
-    password: '123456'
-   }], {});
-
+    await queryInterface.bulkInsert('Users', 
+      [
+        {
+          name: 'Ihor',
+          email: 'siafin2010@gmail.com',
+          password: await bcrypt.hash('123456', SALTROUNDS),
+          isAdmin: true
+        },
+        {
+          name: 'admin',
+          email: 'admin@example.com',
+          password: await bcrypt.hash('passwordsecret', SALTROUNDS),
+          isAdmin: true
+        }
+      ], {});
   },
 
   down: async (queryInterface, Sequelize) => {
@@ -26,7 +39,7 @@ module.exports = {
      * Example:
      * await queryInterface.bulkDelete('People', null, {});
      */
-    await queryInterface.bulkDelete('Users', {email: 'siafin2010@gmail.com'}, {});
+    await queryInterface.bulkDelete('Users', {email: ['siafin2010@gmail.com','admin@example.com']}, {});
 
   }
 };
